Move field io to new submodule
