+++ discretisedfield/util/util.py
@@ -5,8 +5,6 @@ import colorsys
 import numpy as np
 import ubermagutil.units as uu
 
-import discretisedfield as df
-
 axesdict = collections.OrderedDict(x=0, y=1, z=2)
 raxesdict = {value: key for key, value in axesdict.items()}
 
@@ -124,68 +122,6 @@ def hls2rgb(hue, lightness=None, saturation=None, lightness_clim=None):
     )
 
     return rgb.squeeze()
-
-
-def fromvtk_legacy(filename):
-    """Read the field from a VTK file (legacy).
-
-    This method reads vtk files written with discretisedfield <= 0.61.0
-    in which the data is stored as point data instead of cell data.
-    """
-    with open(filename, "r") as f:
-        content = f.read()
-    lines = content.split("\n")
-
-    # Determine the dimension of the field.
-    if "VECTORS" in content:
-        dim = 3
-        data_marker = "VECTORS"
-        skip = 0  # after how many lines data starts after marker
-    else:
-        dim = 1
-        data_marker = "SCALARS"
-        skip = 1
-
-    # Extract the metadata
-    mdatalist = ["X_COORDINATES", "Y_COORDINATES", "Z_COORDINATES"]
-    n = []
-    cell = []
-    origin = []
-    for i, line in enumerate(lines):
-        for mdatum in mdatalist:
-            if mdatum in line:
-                n.append(int(line.split()[1]))
-                coordinates = list(map(float, lines[i + 1].split()))
-                origin.append(coordinates[0])
-                if len(coordinates) > 1:
-                    cell.append(coordinates[1] - coordinates[0])
-                else:
-                    # If only one cell exists, 1nm cell is used by default.
-                    cell.append(1e-9)
-
-    # Create objects from metadata info
-    p1 = np.subtract(origin, np.multiply(cell, 0.5))
-    p2 = np.add(p1, np.multiply(n, cell))
-    mesh = df.Mesh(region=df.Region(p1=p1, p2=p2), n=n)
-    field = df.Field(mesh, dim=dim)
-
-    # Find where data starts.
-    for i, line in enumerate(lines):
-        if line.startswith(data_marker):
-            start_index = i
-            break
-
-    # Extract data.
-    for i, line in zip(mesh.indices, lines[start_index + skip + 1 :]):
-        if not line[0].isalpha():
-            field.array[i] = list(map(float, line.split()))
-
-    return field
-
-
-def strip_extension(filename):
-    """Strip the file extension from a file name (which can be a full path)."""
-    return "".join(filename.split(".")[:-1])
 
 
 def rescale_xarray(array, multiplier):
